Add Virtual Attributes from the User Name parameter
To make the application more personal, we wanted to display the user's first name when they login.
Because users enter their full names in the field during registration, we created 2 functions that split the name parameter into 2 virtual attributes.
The methods are called .first_name and .last_name and they are housed in the User model.

Relates to Ticket: BEDT-106