Add Session Tracker for Dynamic Source Links
We want to monitor when people come from various social media outlets into our portfolio site and eventually we will showcase to them our other available pages.
Right now we've added a method that saves the source from a query string into a session cookie called 'source'.
The method is in Application Controller. We will refactor it in future commits.
We then check to see if there is a source session and if yes, Thank the user for visiting from that source.

Relates to Ticket: BEDT-110