Separate SetSource from ApplicationController
We created the SetSource process that evaluates the source of a visitor and thanks them for visiting.
However this was done in the Application Controller thus breaking the Single Responsibility Principle.

We now migrated that code to be it's own controller concern called SetSource. Which works as a module and is a mixin of the Application Controller.

Relates to Ticket: BEDT-111