Refactor strong parameters in Portfolio items
During creating of the Portfolio Controller, the create and update actions worked on the same parameters and we repeated them.
This led to breaking of the DRY principle.

We have refactored this into a method that now holds the strong parameters as a private method which any of the controller methods can call.

Relates to Ticket: BEDT-112