Add Guest User as a Current User
Because of the NullObject Pattern principle, we added an ostruct as a guest user to allow the system to recognize a guest visitor and a signed in visitor.
All of them are tied to the CurrentUser object from the Devise Controller thus we also had to use a class distinguisher in order to show the right links to show.

Also refactored all concerns to have a concern suffix for code readability.
The current user is maintained as a controller concern as a mixin of the appliaction controller.

Relates to Ticket: BEDT-114