Add Custom layout for Blog and Portfolio
We added 2 unique layouts for the blog and the portfolio in order to allow each page to load differently from the other pages.

the new layouts are in the layouts folder as 'blog.html.erb' and 'portfolio.html.erb'
We also notified the controller to specifically load their individual stylesheets.

Relates to Ticket: BEDT-118