Implement Partials with data for Portfolios and Nav
Initial form for creating or editing a new form was implemented separately.
This led to duplication and breaking of the DRY principle.
We now refactored the form as a partial '_form.html.erb' and rendered it for both the new and edit views.

For the nav Partial, we also updated it to be able to take dynamic data for classes in order to allow unique styling.
the variable on the Partial is 'nav_location'.

Relates to Ticket: BEDT-120