Add login helper for application, blog & portfolio views
The authentication links(register, login, logout) were implemented individually on each main view.
This led to duplication of code and breaching the DRY principle.

Considering majority of the auth link code were conditional and pure ruby code, we opted to implement it as a helper method in the application helper module.

Related to JIRA Ticket: BEDT-121