Refactor Session Trackers to Session Source View Helper method
Initially we added our session trackers for each layout independently.
This led to duplication of code and breaking the DRY principle.

We then recreated the session source trackers as a view helper method -'session_source_helper' in the application helper module.
It also uses the content_tag to generate the html tag and assigns it a CSS class - 'session-source-greeting'

Relates to JIRA Ticket: BEDT-122