Add custom formatting using the ActionView Helper methods
We added the Created Date field in the blog view.
By default it displays the date as a timestamp as how it is saved in the db.
To make it more user friendly, we used the '.distance_of_time_in_words' helper method to calculate the time since creation date and counts down from the current date.

Also in the contact page, we displayed our phone number.
And to make it more user friendly, we used the'.number_to_phone' helper method to format and display the number better.

Relates to JIRA Ticket: BEDT-125