Refactor @Portfolio_items variable in the Portfolio Controller
Initially we were calling the @portfolio_items independently for several actions in the Portfolio controller.
This led to duplication of code and breaking the DRY principle.

Solution was to create a method '.set_portfolio_items' method and ensure that it is run before any other action by calling it in the '.before_action' method
The actions that need this method include : #edit, #show, #update

Relates to JIRA Ticket: BEDT-131