Refactor copyrights generation from an internal Gem and Application helper
Initially we built the whole copyrights generation as a module in the application layout.
We then refactored it to be a separate gem called - MradiViewTool with class renderer and classmethod '.copyright'
We then called it in the app to be able to use it.
Finally we moved the actual generation of the copyright from the application layout to be a View Helper method - 'copyright_generator_helper'.

Relates to JIRA Ticket: BEDT-139