Fix login helper argument error
Because we adjusted the login helper class to have a styles argument, the blog and portfolio layouts that also use the helper started failing because they are not sending any styles.
Thus before working on the layouts, we adjusted the code to have a default so that the application doesn't crash even when a layout doesn't send a style argument.

Relates to JIRA Ticket: BEDT-156