Add styles to Portfolio show page and add 'View All' link to footer
Just added normal grid layout to the show page: container wrapper, 1 row and 2 cols to house the content.

Also added a 'View all' link to help UX for people to be able to easily go to the portfolio index without having to scroll back to the top of the page.

Relates to JIRA Ticket: BEDT-165