Refactor the navigation to a single helper method & Add active indicator for application layout
We migrated all the link generation for each navigation to be within 1 helper method to help reduce code duplication.
The helper method also makes use of an array with the navigation and path stored as hashes to allow flexibility and scaling in future.

We also added another boolen method that returns active to help add the active class to the navigation of the application layout to show an active page.
For the Portfolio and Blog layout though , this is not effective as both layouts don't make use of an active status to show current page.

Relates to JIRA Tickets: BEDT-167, BEDT-168