Add Drag & Drop Interface (FED Only) to change portfolio position
We first added a css class '.sortable' to the Portfolio index view.
We added a function in the Portfolio.js - 'ready' that takes the class and passes it to the .sortable method (from the HTML5sortable library).
This allows us to be able to take an element on the page and move it around and adjust accordingly.

Relates to JIRA Ticket: BEDT-173