Add Persistent positions on drag and drop feature Portfolio index
We added more controls to the application to allow us to have the application save the position after changing it.
In the portfolio.coffee we set a new method - set_position that calculates the position of each card.
Then we updated the ready function to call set function on load and also after changing a card's position.
Finally we used an ajax call to send the data to be saved in the db.
In the controller we added a new action that allowed us to save data sent to it by js file. Also we added 'render nothing: true' to allow the system to not complain if it didn't find a sort view.
We also had to update the routes to allow a custom url to the sort action.

Relates to JIRA Ticket: BEDT-174