Add authorization to the Drag and Drop feature
We are now limiting the Drag and drop feature to only site admins.
This was done by adding sort to the list of except actions for normal users.
and adding a condition to the index view to only add the '.sortable' class when a site_admin user is logged in.

Relates to JIRA Ticket: BEDT-175