Add Full Bg video to application layout
We updated the application layout with a video tag which allows us to add several options to the layout.
A video file - we used one that is less than 5mb for performance, autoplay to be true, loop to be true and muted to be true, lastly we also added a poster to our initial bg image incase the video fails to load.

Relates to JIRA Ticket: BEDT-179