Add Notifications to the app
We installed the Gritter gem to help us with managing alerts/notifications in the system.
We also generated a locale incase we need to add other languages in future.

We created a partial '_alerts' housed in the shared folder in views which contains the logic to display errors.

Relates to JIRA Ticket: BEDT-185