Implement connection to Twitter API
Added the most popular Twitter gem.
Created a separate file called SocialTool in our lib section that will make use of the gem.
There is a class method that is called '.twitter_search' that creates a connection to the API and looks for recent tweets that have the hashtags '#tech & #news'.

Relates to JIRA Ticket: BEDT-188