Add Redis library & Comment generator
We added Redis v4 to the gemfile. This is what Action cable will be using.

Also added the Comment generator. Used the resource generator as we do not need a unique view for Comments.
Also added relations to a user and blog. It belongs to both, and all has many comments.
Also added a dependency :destroy to allow data integrity when removing either a user/blog.

Relates to JIRA Ticket: BEDT-192