Add new comment form, partial for comments and controller actions
Added a form where logged in users will be able to add comments in the show view page.
Also added a partial '_comment' in the Comment view that will be loaded in the blog show page to show entered view pages.

Integrated controller actions to allow us to reference comments in the blog show view page.
Also overrode the set blog for show page to have the '.include' method to improve performance when calling comments.

Relates to JIRA Ticket: BEDT-193