Add redirect to blogs path when non-admin accesses draft posts
Added an if-else condition to check if user is an admin or the post is published in the blogs_controller/show action before allowing access to a page.
When a non-admin tries to click a draft post he is redirected to blogs path and notified they are not allowed to view it.

Relates to JIRA Ticket: BEDT-200