Add topic dropdown form element and validation
We added a new form element (collection_select) which allows dropdowns in a form.
We made it to always select the current topic id in the edit page and included a blank option.

We also added validation to the blogs_model to ensure someone has to select a topic id for data integrity.

Relates to JIRA Ticket: BEDT-201