Add a blog widget in the sidebar that loads only topics that have blog posts
We added a new scope that filters topics that only have blog posts.
We then added a widget to the sidebar that utilizes an instance variable from the scope to load the topics.
We then added a private method to both blogs_controller and topics_controller that loads the instance variable.
The above technically breaks the DRY principle with the exception where in the blogs_controller there's variation on where it should be loaded.

Relates to JIRA Ticket: BEDT-203