Update styles for Blog Show page & Comment form
Updated the show page for blog to have more spacing between the links and body.
Made the topic title a badge.
Added a custom label for the custom form for someone to know that's separate from the post. Made the font weight of the label more bolder to make it more distinct.